### 2024-09-10 #7215 @editorialbot
[PRE REVIEW]: ESAT: Environmental Source Apportionment Toolkit Python package
**Submitting author:** @deronsmith (<a href="http://orcid.org/0009-0009-4015-5270">Deron Smith</a>)
**Repository:** https://github.com/quanted/esat
**Branch with paper.md** (empty if default branch):
**Version:** 2024.0.2
