### 2024-09-12 4ae3bb7
paper ref doi updates, paper word count reduced
