+++ paper/paper.bib
@@ -18,7 +18,8 @@
   pages={626--635},
   year={2015},
   publisher={Elsevier},
-  url={https://doi.org/10.1016/j.scitotenv.2015.01.022}
+  url={https://doi.org/10.1016/j.scitotenv.2015.01.022},
+  doi={10.1016/j.scitotenv.2015.01.022}
 }
 
 @article{DeMelo:2012,
@@ -36,7 +37,8 @@
   pages={45--55},
   year={2008},
   publisher={IEEE},
-  url={https://doi.org/10.1109/TPAMI.2008.277}
+  url={https://doi.org/10.1109/TPAMI.2008.277},
+  doi={10.1109/TPAMI.2008.277}
 }
 
 @article{Jiang:2019,
@@ -88,7 +90,8 @@
   pages={854--888},
   year={1999},
   publisher={Taylor \& Francis},
-  url={https://doi.org/10.1080/10618600.1999.10474853}
+  url={https://doi.org/10.1080/10618600.1999.10474853},
+  doi={10.1080/10618600.1999.10474853}
 }
 
 @article{Paatero:2014,
@@ -118,6 +121,7 @@
   pages={1--10},
   year={2006},
   publisher={Springer},
-  url={https://doi.org/10.1186/1471-2105-7-175}
+  url={https://doi.org/10.1186/1471-2105-7-175},
+  doi={10.1186/1471-2105-7-175}
 }
 
