+++ paper/paper.md
@@ -35,40 +35,38 @@ bibliography: paper.bib
 
 Source apportionment is an important tool in environmental science where sample or sensor data are often the product
 of many, often unknown, contributing sources. Source apportionment is used to understand the relative contributions of 
-air sources [@Bhandari:2022] like vehicle emissions, industrial activities, biomass burning, dust. As well as particulate matter 
-pollution and to identify relative contributions of point sources (e.g., wastewater treatment discharges) and 
-non-point sources (e.g., agricultural runoff) in water bodies such as lakes, rivers, and estuaries [@Mamum:2021] [@Jiang:2019]. Using non-negative 
-matrix factorization(NMF), source apportionment models estimate potential source profiles and contributions providing a 
-cost-efficient method for further strategic data collection or modeling. 
-
+air sources [@Bhandari:2022] like vehicle emissions, industrial activities, or dust. As well as particulate matter 
+pollution and to identify relative contributions of point sources and non-point sources in water bodies such as lakes, 
+rivers, and estuaries [@Mamum:2021] [@Jiang:2019]. Using non-negative matrix factorization(NMF), source apportionment 
+models estimate potential source profiles and contributions providing a cost-efficient method for further strategic data 
+collection or modeling.
 
 Environmental Source Apportionment Toolkit (ESAT) is an open-source Python package that provides a flexible and 
 transparent workflow for source apportionment using NMF algorithms, developed to replace the EPA's Positive 
-Matrix Factorization version 5 (PMF5) application [@PMF5:2014] [@Paatero:1999]. `ESAT` recreates the source apportionment workflow of 
-PMF5 including pre-post processing analytical tools, batch modeling, model uncertainty estimations and customized 
-constraints. Additionally, `ESAT` offers a simulator for generating datasets from synthetic profiles and contributions, 
-allowing for direct model output evaluation. The synthetic profiles can either be randomly generated, use a pre-defined 
-set of profiles, or be a combination of the two. The random synthetic contributions can follow specified curves and 
-value ranges. Running `ESAT` using the synthetic datasets one is able to accurately assess ESAT's ability to find a 
-solution that recreates the original synthetic profiles and contributions. 
+Matrix Factorization version 5 (PMF5) application [@PMF5:2014] [@Paatero:1999]. `ESAT` recreates the source apportionment 
+workflow of PMF5 including pre-post processing analytical tools, batch modeling, uncertainty estimations and customized 
+constraints. `ESAT` offers a simulator for generating datasets from synthetic profiles and contributions, allowing for 
+model output evaluation. The synthetic profiles can be randomly generated, use a pre-defined set of profiles, or be a 
+combination of the two. The random synthetic contributions can follow specified curves and value ranges. Running `ESAT` 
+using the synthetic datasets one is able to accurately assess ESAT's ability to find a solution that recreates the 
+original synthetic profiles and contributions. 
 
 # Statement of Need
 
-`ESAT` has been developed as a replacement to PMF5, and has been designed for increased flexibility, documentation and 
-transparency. The EPA's PMF5, released in 2014, provides a widely-used source apportionment modeling and analysis 
-workflow that is no longer supported and relies on the proprietary Multilinear Engine v2 (ME2) that lacks documentation.
+The EPA's PMF5, released in 2014, provides a widely-used source apportionment modeling and analysis 
+workflow that is no longer supported and relies on the proprietary Multilinear Engine v2 (ME2). `ESAT` has been 
+developed as a replacement to PMF5, and has been designed for increased flexibility, documentation and transparency. 
 
-The Python API and CLI of `ESAT` provides an object-oriented interface that can recreates the PMF5 workflow. 
[…]
