### 2024-09-12 #7215 @deronsmith
Hi @kthyng, thank you for this information. I have gone ahead and addressed the missing doi and reduced the paper word count to under 1000 (pushed to the main branch). I will review the database of reviewers for potential options and get back to you with that information.
