### 2024-09-12 #7215 @kthyng
@deronsmith
* have you thought about putting your notebooks into your docs? That would be nice for users!
* I see PyPI is in your plan for the future — would be great to have that as part of the JOSS review process.
