### 2024-09-16 #7215 @deronsmith
> @deronsmith
>
> * have you thought about putting your notebooks into your docs? That would be nice for users!
> * I see PyPI is in your plan for the future — would be great to have that as part of the JOSS review process.

Absolutely, I'll move a few of the more polished workflow notebooks into the docs.

The EPA's QA policy requires an external peer review prior to code or software release, so we were planning on pushing to pypi once we have the JOSS review completed (assuming we successfully publish through JOSS).  But I believe the prebuilt wheels are available to logged in Github user's by looking at the 'Build and Publish Wheel' workflow.
