### 2024-09-16 #7215 @deronsmith
I had actually linked to the two primary notebooks in the README, which is also shown in the documentation readme page https://quanted.github.io/esat/README.html
