+++ docs/html/_sources/docs/esat.rotational.rst
@@ -1,0 +1,21 @@
+esat.rotational package
+=======================
+
+Submodules
+----------
+
+esat.rotational.constrained module
+----------------------------------
+
+.. automodule:: esat.rotational.constrained
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: esat.rotational
+   :members:
+   :undoc-members:
+   :show-inheritance:
