+++ docs/html/_sources/paper/paper.md
@@ -1,0 +1,121 @@
+---
+title: 'ESAT: Environmental Source Apportionment Toolkit Python package'
+tags: 
+    - Python
+    - Rust
+    - source apportionment
+    - matrix factorization
+    - environmental research
+    - air quality
+    - water quality
+authors:
+    - name: Deron Smith
+      orcid: 0009-0009-4015-5270
+      affiliation: 1
+    - name: Michael Cyterski
+      orcid: 0000-0002-8630-873X
+      affiliation: 1
+    - name: John M Johnston
+      orcid: 0000-0002-5886-7876
+      affiliation: 1
+    - name: Kurt Wolfe
+      orcid: 
+      affiliation: 1
+    - name: Rajbir Parmar
+      orcid: 0009-0005-2221-0433
+      affiliation: 1
+affiliations:
+    - name: United States Environmental Protection Agency, Office of Research and Development, Center for Environmental Measurement and Modeling
+      index: 1
+date: 9 September 2024
+bibliography: paper.bib
+---
+
+# Summary
+
+Source apportionment is an important tool in environmental science where sample or sensor data are often the product
+of many, often unknown, contributing sources. Source apportionment is used to understand the relative contributions of 
+air sources [@Bhandari:2022] like vehicle emissions, industrial activities, or dust. As well as particulate matter 
+pollution and to identify relative contributions of point sources and non-point sources in water bodies such as lakes, 
+rivers, and estuaries [@Mamum:2021] [@Jiang:2019]. Using non-negative matrix factorization(NMF), source apportionment 
+models estimate potential source profiles and contributions providing a cost-efficient method for further strategic data 
+collection or modeling.
+
+Environmental Source Apportionment Toolkit (ESAT) is an open-source Python package that provides a flexible and 
+transparent workflow for source apportionment using NMF algorithms, developed to replace the EPA's Positive 
+Matrix Factorization version 5 (PMF5) application [@PMF5:2014] [@Paatero:1999]. `ESAT` recreates the source apportionment 
+workflow of PMF5 including pre-post processing analytical tools, batch modeling, uncertainty estimations and customized 
+constraints. `ESAT` offers a simulator for generating datasets from synthetic profiles and contributions, allowing for 
+model output evaluation. The synthetic profiles can be randomly generated, use a pre-defined set of profiles, or be a 
+combination of the two. The random synthetic contributions can follow specified curves and value ranges. Running `ESAT` 
+using the synthetic datasets one is able to accurately assess ESAT's ability to find a solution that recreates the 
+original synthetic profiles and contributions. 
+
+# Statement of Need
+
+The EPA's PMF5, released in 2014, provides a widely-used source apportionment modeling and analysis 
+workflow that is no longer supported and relies on the proprietary Multilinear Engine v2 (ME2). `ESAT` has been 
+developed as a replacement to PMF5, and has been designed for increased flexibility, documentation and transparency. 
+
+The Python API and CLI of `ESAT` provides a programmatic interface that can recreate the PMF5 workflow. 
+The matrix factorization algorithms in `ESAT` have been written in Rust for runtime optimization. The `ESAT` API and CLI
+provides a flexible way to create source apportionment workflows and novel research applications. `ESAT` was developed for 
+environmental research, though it's not limited to that domain, as matrix factorization is used in many different fields.
+
+## Algorithms
[…]
