+++ docs/html/_sources/docs/modules.rst
@@ -1,0 +1,12 @@
+Python API
+===============
+
+.. toctree::
+   :maxdepth: 3
+
+   esat.rst
+   esat.data.rst
+   esat.error.rst
+   esat.model.rst
+   esat.rotational.rst
+   eval.rst
