+++ requirements.txt
@@ -6,4 +6,4 @@ plotly==5.19.0
 scipy==1.12.0
 tabulate==0.8.10
 tqdm==4.66.2
-typer==0.4.2
+typer==0.9.4
