### 2024-09-24 0af7096
updated sphinx configuration to resolve warnings
