+++ docs/_sources/docs/esat.data.rst
@@ -1,0 +1,37 @@
+esat.data package
+=================
+
+Submodules
+----------
+
+esat.data.analysis module
+-------------------------
+
+.. automodule:: esat.data.analysis
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.data.datahandler module
+----------------------------
+
+.. automodule:: esat.data.datahandler
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.data.test\_tools module
+----------------------------
+
+.. automodule:: esat.data.test_tools
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: esat.data
+   :members:
+   :undoc-members:
+   :show-inheritance:
