+++ docs/html/_sources/index.rst
@@ -3,8 +3,8 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-Welcome to EPA's ESAT Documentation
-=================================
+EPA's ESAT Documentation
+========================
 
 .. toctree::
    :maxdepth: 3
