### 2024-09-24 0383f98
updated sphinx configuration to resolve warnings
