+++ conf.py
@@ -35,7 +35,7 @@ autodoc_default_options = {
 }
 
 templates_path = ['_templates']
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', ".pytest_cache", "paper"]
 
 
 # -- Options for HTML output -------------------------------------------------
@@ -50,3 +50,6 @@ html_sidebars = {
         'globaltoc.html'
     ]
 }
+
+# myst_enable_extensions = ["deflist"]
+myst_heading_anchors = 3
