### 2024-09-26 #7215 @mengqi-z
👋@Zeitsperre @nanhung @katmratliff @ZiyiHuang0708 - Would any of you be willing o review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
