### 2024-10-02 #7215 @mengqi-z
👋 @nickrsan @ifoxfoot - Would you be willing to review this submission for JOSS? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
