### 2024-10-04 #7215 @mengqi-z
👋 @ibarraespinosa @purushottamkar @niravlekinwala - Would any of you be interested in reviewing this submission for JOSS?
JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
