### 2024-10-04 #7316 @editorialbot
[REVIEW]: ESAT: Environmental Source Apportionment Toolkit Python package
**Submitting author:** @deronsmith (<a href="http://orcid.org/0009-0009-4015-5270">Deron Smith</a>)
**Repository:** https://github.com/quanted/esat
**Branch with paper.md** (empty if default branch):
**Version:** v2024.0.3-alpha
**Archive:** 10.23719/1531870
