### 2024-10-04 quanted/esat#6 @ifoxfoot
Need year and copywrite holder in license
https://github.com/openjournals/joss-reviews/issues/7316

https://github.com/quanted/esat/blob/0af70969baae5500daebab1b42626b9cc93a762e/LICENSE#L189

Hi! I think you need to put the copywrite holder and the year into the license text.
