### 2024-10-04 #7316 @mengqi-z
@openjournals/dev - The PDF is not generating correctly. Could you please help me looking into the issue? Thank you!

@gutabeshu, @ifoxfoot, @niravlekinwala @deronsmith - In the meantime, you can access the most recent PDF from the Pre-Review stage here: https://github.com/openjournals/joss-reviews/issues/7215#issuecomment-2347314129. Apologies for the inconvenience!
