### 2024-10-05 #7316 @arfon
@mengqi-z – the issue here is that there are two `paper.md` files in the repository and there must only be one. @deronsmith – could you clean this up please by removing or renaming the one that _is not_ the JOSS paper?
