### 2024-10-05 #7316 @mengqi-z
@arfon - Ah, I see! It turns out there is a `paper.md` file located under `esat/docs/html/_sources/paper/`. Thanks for spotting the issue!
