+++ docs/html/_sources/docs/esat.rotational.rst.txt
@@ -1,21 +1,0 @@
-esat.rotational package
-=======================
-
-Submodules
-----------
-
-esat.rotational.constrained module
-----------------------------------
-
-.. automodule:: esat.rotational.constrained
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: esat.rotational
-   :members:
-   :undoc-members:
-   :show-inheritance:
