+++ docs/html/_sources/docs/esat.model.rst.txt
@@ -1,61 +1,0 @@
-esat.model package
-==================
-
-Submodules
-----------
-
-esat.model.batch\_sa module
----------------------------
-
-.. automodule:: esat.model.batch_sa
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.ls\_nmf module
--------------------------
-
-.. automodule:: esat.model.ls_nmf
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.optimization module
-------------------------------
-
-.. automodule:: esat.model.optimization
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.recombinator module
-------------------------------
-
-.. automodule:: esat.model.recombinator
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.sa module
---------------------
-
-.. automodule:: esat.model.sa
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.ws\_nmf module
--------------------------
-
-.. automodule:: esat.model.ws_nmf
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: esat.model
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
