+++ docs/html/_sources/docs/src.model.rst.txt
@@ -1,53 +1,0 @@
-src.model package
-=================
-
-Submodules
-----------
-
-src.model.batch\_sa module
---------------------------
-
-.. automodule:: src.model.batch_sa
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.model.ls\_nmf module
-------------------------
-
-.. automodule:: src.model.ls_nmf
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.model.optimization module
------------------------------
-
-.. automodule:: src.model.optimization
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.model.recombinator module
------------------------------
-
-.. automodule:: src.model.recombinator
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.model.sa module
--------------------
-
-.. automodule:: src.model.sa
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-src.model.ws\_nmf module
-------------------------
-
-.. automodule:: src.model.ws_nmf
-   :members:
-   :undoc-members:
-   :show-inheritance:
