+++ docs/html/_sources/docs/esat.cli.rst
@@ -1,21 +1,0 @@
-esat.cli package
-================
-
-Submodules
-----------
-
-esat.cli.esat\_cli module
--------------------------
-
-.. automodule:: esat.cli.esat_cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: esat.cli
-   :members:
-   :undoc-members:
-   :show-inheritance:
