+++ docs/_sources/docs/esat.model.rst
@@ -1,53 +1,0 @@
-esat.model package
-==================
-
-Submodules
-----------
-
-esat.model.batch\_sa module
----------------------------
-
-.. automodule:: esat.model.batch_sa
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.ls\_nmf module
--------------------------
-
-.. automodule:: esat.model.ls_nmf
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.recombinator module
-------------------------------
-
-.. automodule:: esat.model.recombinator
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.sa module
---------------------
-
-.. automodule:: esat.model.sa
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-esat.model.ws\_nmf module
--------------------------
-
-.. automodule:: esat.model.ws_nmf
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: esat.model
-   :members:
-   :undoc-members:
-   :show-inheritance:

[…]
