### 2024-10-08 #7316 @deronsmith
> @arfon - Ah, I see! It turns out there is a `paper.md` file located under `esat/docs/html/_sources/paper/`. Thanks for spotting the issue!

Thank you for your patience while this issue was resolved. The paper.md file in docs has been removed and the corresponding directories added to gitignore to prevent future issues.
