### 2024-10-08 quanted/esat#7 @ifoxfoot
References and minor quality of writing issue with JOSS paper
https://github.com/openjournals/joss-reviews/issues/7316

References:
Hello. Your citations look great, except when you have two next to each other (e.g. `[@Mamum:2021] [@Jiang:2019]` ) you should separate them with a semicolon. Like this `[@Mamum:2021; @Jiang:2019]`

Quality of Writing:
https://github.com/quanted/esat/blob/7a40dfdbecfcb52ca5c724295ed2d7dae5049a44/paper/paper.md?plain=1#L37C47-L40C51
I think this was meant to be one sentence?

https://github.com/quanted/esat/blob/7a40dfdbecfcb52ca5c724295ed2d7dae5049a44/paper/paper.md?plain=1#L40C78-L40C96
You're missing a space here
