### 2024-10-08 45b219a
citation formatting and paper update to address issue 7
