+++ paper/paper.md
@@ -35,15 +35,15 @@ bibliography: paper.bib
 
 Source apportionment is an important tool in environmental science where sample or sensor data are often the product
 of many, often unknown, contributing sources. Source apportionment is used to understand the relative contributions of 
-air sources [@Bhandari:2022] like vehicle emissions, industrial activities, or dust. As well as particulate matter 
+air sources [@Bhandari:2022] like vehicle emissions, industrial activities, or dust; as well as particulate matter 
 pollution and to identify relative contributions of point sources and non-point sources in water bodies such as lakes, 
-rivers, and estuaries [@Mamum:2021] [@Jiang:2019]. Using non-negative matrix factorization(NMF), source apportionment 
+rivers, and estuaries [@Mamum:2021; @Jiang:2019]. Using non-negative matrix factorization (NMF), source apportionment 
 models estimate potential source profiles and contributions providing a cost-efficient method for further strategic data 
 collection or modeling.
 
 Environmental Source Apportionment Toolkit (ESAT) is an open-source Python package that provides a flexible and 
 transparent workflow for source apportionment using NMF algorithms, developed to replace the EPA's Positive 
-Matrix Factorization version 5 (PMF5) application [@PMF5:2014] [@Paatero:1999]. `ESAT` recreates the source apportionment 
+Matrix Factorization version 5 (PMF5) application [@PMF5:2014; @Paatero:1999]. `ESAT` recreates the source apportionment 
 workflow of PMF5 including pre-post processing analytical tools, batch modeling, uncertainty estimations and customized 
 constraints. `ESAT` offers a simulator for generating datasets from synthetic profiles and contributions, allowing for 
 model output evaluation. The synthetic profiles can be randomly generated, use a pre-defined set of profiles, or be a 
@@ -68,7 +68,7 @@ the product of a factor contribution matrix (W) and a factor profile matrix (H),
 [@Paatero:1994]. The goal is to find factor matrices that best reproduce the input matrix, while constraining all, 
 or most of, the factor elements to be non-negative. The solution, W and H, can be used to calculate the 
 residuals and overall model loss. `ESAT` has two NMF algorithms for updating the profile and contribution matrices: 
-least-squares NMF (LS-NMF) [@Wang:2006] and weighted-semi NMF (WS-NMF) [@Ding:2008] [@DeMelo:2012]. 
+least-squares NMF (LS-NMF) [@Wang:2006] and weighted-semi NMF (WS-NMF) [@Ding:2008; @DeMelo:2012]. 
 
 The loss function used in `ESAT`, and PMF5, is a variation of squared-error loss, where data uncertainty is taken into
 consideration (both in the loss function and in the matrix update equations):
@@ -97,7 +97,7 @@ where $W^{-} = \frac{(|W| - W)}{2.0}$ and $W^{+} = \frac{(|W| + W)}{2.0}$.
 
 ## Error Estimation
 An important part of the source apportionment workflow is quantifying potential model error. `ESAT` offers the error 
-estimation methods that were developed and made available in PMF5 [@Brown:2015] [@Paatero:2014].
+estimation methods that were developed and made available in PMF5 [@Brown:2015; @Paatero:2014].
 
 The displacement method (DISP) determines the amount that a source profile feature, a single value in the H matrix, 
 must increase and decrease to cause targeted changes to the loss value. One or more features can be selected
