### 2024-10-08 quanted/esat#7 @deronsmith
Thank you for your feedback. All issue items have been addressed in commit 45b219a.
