### 2024-10-16 40f85a4
added k-estimate methods bootstrap and input pertubation
