### 2024-10-22 quanted/esat#9 @ifoxfoot
Help installing ESAT with whl files
https://github.com/openjournals/joss-reviews/issues/7316

Hello,

I'm having trouble installing ESAT using the installation instructions in the readme. I'm using Windows 10 Enterprise, 64-bit OS. I'm using a conda environment with python 3.12.7

I've tried running `pip install esat-2024.0.2-cp312-cp312-win_amd64.whl` in an empty project and in a project with a cloned version of esat and I get the error
```
WARNING: Requirement 'esat-2024.0.2-cp312-cp312-win_amd64.whl' looks like a filename, but the file does not exist
Processing c:\users\u4eewirf\desktop\esat\esat-2024.0.2-cp312-cp312-win_amd64.whl
ERROR: Could not install packages due to an OSError: [Errno 2] No such file or
directory: 'C:\\Users\\u4eewirf\\Desktop\\esat\\esat-2024.0.2-cp312-cp312-win_amd64.whl'
```
I went to your github actions and downloaded the artifact for the wheel I wanted, unzipped it, and put it in the folder with the cloned version of esat and the installation worked. I used the `esat-2024.0.2-cp312-cp312-win_amd64.whl` wheel. Based on the installation instructions in the readme, it seems like the wheel file is supposed to be in the repo somewhere?

I also think that picking a code version is somewhat confusing because there aren't any code releases. Maybe you could include a list of recent code versions available to install in the readme?
