### 2024-10-23 quanted/esat#9 @deronsmith
Hey,

I apologize for the confusion. Once we have an external review (this JOSS review), we will be allowed to deploy the package to pypi and that will be the primary method for users to install the package. Though I can go ahead and include the necessary steps to install from the github actions, such as you completed, in the README.

Only the latest version is being compiled by the github actions and we will begin to increase the version number moving forward (though it will be specific updates, fixes not for all github pushes).
