### 2024-10-23 035d5f9
updates to the readme for steps to install ESAT using the github action artifacts
