### 2024-10-23 quanted/esat#6 @deronsmith
Added to LICENSE in commit f8c8dcf
