+++ _sources/docs/esat.error.rst
@@ -1,0 +1,45 @@
+esat.error package
+==================
+
+Submodules
+----------
+
+esat.error.bootstrap module
+---------------------------
+
+.. automodule:: esat.error.bootstrap
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.error.bs\_disp module
+--------------------------
+
+.. automodule:: esat.error.bs_disp
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.error.displacement module
+------------------------------
+
+.. automodule:: esat.error.displacement
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.error.error module
+-----------------------
+
+.. automodule:: esat.error.error
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: esat.error
+   :members:
+   :undoc-members:
+   :show-inheritance:
