### 2024-10-23 quanted/esat#9 @ifoxfoot
Hi @deronsmith, good to hear the package will be available on pypi, I think that will make installation a lot easier! For now I do think you should have installation instructions in the README, it will help with the JOSS reviewer checklist. I see you've already made some changes to the README so I will close this issue!
