+++ esat/data/datahandler.py
@@ -271,11 +271,12 @@ class DataHandler:
         if input_data is None:
             input_data = self.input_data.to_numpy()
         try:
-            optimal_blocks = optimal_block_length(input_data)
-            optimal_block = []
-            for opt in optimal_blocks:
-                optimal_block.append(opt.b_star_cb)
-            self.optimal_block = math.floor(np.mean(optimal_block))
+            with np.errstate(divide='ignore', invalid='ignore'):
+                optimal_blocks = optimal_block_length(input_data)
+                optimal_block = []
+                for opt in optimal_blocks:
+                    optimal_block.append(opt.b_star_cb)
+                self.optimal_block = math.floor(np.mean(optimal_block))
         except ValueError as e:
             self.optimal_block = int(input_data.shape[1]/5)
             logger.error(f"Unable to determine optimal block size. Setting default to {self.optimal_block}")
