### 2024-11-05 8976c17
corrected test for H initialization with prior H
