### 2024-11-05 e856391
corrected test for H initialization with prior H
