### 2024-11-05 2acf28e
corrected test for H initialization with prior H
