### 2024-11-05 1d688db
corrected test for H initialization with prior H
