+++ data/test_output/cli_test/sim_config.toml
@@ -1,29 +1,0 @@
-[project]
-directory = D:\git\esat\tests\cli\..\..\data\test_output\cli_test
-
-[data]
-input_path = synthetic_data.csv
-uncertainty_path = synthetic_uncertainty.csv
-
-[parameters]
-seed = 42
-factors_n = 6
-features_n = 40
-samples_n = 200
-outliers = True
-outlier_p = 0.1
-outlier_mag = 2
-contribution_max = 10
-noise_mean_min = 0.03
-noise_mean_max = 0.05
-noise_scale = 0.02
-uncertainty_mean_min = 0.04
-uncertainty_mean_max = 0.06
-uncertainty_scale = 0.01
-
-[contributions]
-0 = {"curve_type":"logistic","scale": 0.1,"frequency":0.5}
-1 = {"curve_type":"periodic","minimum": 0.1,"maximum": 0.9,"frequency": 0.5,"scale":0.1}
-2 = {"curve_type":"increasing","minimum": 0.1,"maximum": 0.9,"scale":0.1}
-3 = {"curve_type":"decreasing","minimum": 0.1,"maximum": 0.9,"scale":0.1}
-
