+++ data/test_output/cli_test/error_config.toml
@@ -1,24 +1,0 @@
-[project]
-project_config = D:\git\esat\tests\cli\..\..\data\test_output\cli_test\run_config.toml
-constrained_config = 
-selected_model = 1
-
-[disp]
-features = [0]
-
-[bs]
-bootstrap_n = 2
-block_size = -1
-threshold = 0.6
-seed = -1
-
-[bsdisp]
-bootstrap_output = 
-bootstrap_n = 20
-block_size = -1
-threshold = 0.6
-seed = -1
-threshold_dq = 0.1
-max_search = 50
-features = []
-
