### 2024-11-10 #7316 @niravlekinwala
@deronsmith @mengqi-z

I have completed my review. I have gone through the paper and example Jupyter notebooks and the code runs without a glitch (except for when factors are 2 & 3, @deronsmith has mentioned it as a bug on the repository page). I have raised a couple of issues on the repository's page as a suggestions for @deronsmith.

Also @mengqi-z, I have completed the checklist.
