### 2024-11-18 quanted/esat#12 @mengqi-z
JOSS Paper Proofread
Hi @deronsmith,

Thank you for submitting your work to JOSS. This issue corresponds to the JOSS paper under review: https://github.com/openjournals/joss-reviews/issues/7316

I have a few minor suggestions to improve the readability of the paper. The line numbers are based on the latest PDF rendered here https://github.com/openjournals/joss-reviews/issues/7316#issuecomment-2483812709

1. Line 21: Consider replacing "pre-post processing" with "pre- and post-processing".
2. Line 26: Consider rephrasing "Running `ESAT` using the synthetic datasets one is able to ..." to "By running `ESAT` using the synthetic datasets, users are able to ..." for smoother readability.
3. Line 26: Please use code style (`ESAT`) for the term "ESAT" to align with tool references.
4. Line 52: Please double check whether the phrase "samples by factors=k" should be revised to "samples by factors=K" for consistency with the description (e.g., M and N) used in Line 50.
5. Line 79: Please use "U.S. EPA" for consistency with the phrasing in Line 76.

Thanks!
