### 2024-11-18 8c6f06c
updates to paper to address issues raised in #12
