### 2024-11-18 b090911
updates to paper to address issues raised in #12
