### 2024-11-18 quanted/esat#12 @deronsmith
Thank you for the feedback, all of these items have been addressed in b090911 .
