### 2024-11-18 #7316 @mengqi-z
@deronsmith - I have completed proofreading the paper and all is good. Thank you.

At this point, could you:

- [ ] Double check authors and affiliations (including ORCIDs).
- [ ] Make a release of the software with the latest changes from the review and **post the version number here**. This is the version that will be used in the JOSS paper.
- [ ] Archive the released software on Zenodo or a similar service (e.g., figshare, an institutional repository).
- [ ] Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
- [ ] **Please post the DOI of the archived version here**.
