### 2024-11-19 8a8bf1b
updated author orcids, package version and added 3.13 to github action list
