### 2024-11-19 15822d1
updated author orcids, package version and added 3.13 to github action list
