+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "esat"
-version = "2024.0.2"
+version = "2024.0.3"
 requires-python = ">=3.10"
 description = "EPA's Environmental Source Apportionment Toolkit python package."
 authors = [
