+++ .github/workflows/python-package.yml
@@ -76,7 +76,7 @@ jobs:
 #      Publish Python 🐍 distribution 📦 to PyPI
 #    if: startsWith(github.ref, 'refs/tags/')  # only publish to PyPI on tag pushes
 #    needs:
-#      - build
+#      - build_wheels
 #      - build_windows_wheel
 #    runs-on: ubuntu-latest
 #    environment:
@@ -140,7 +140,8 @@ jobs:
   publish-to-testpypi:
     name: Publish Python 🐍 distribution 📦 to TestPyPI
     needs:
-      - build
+      - build_wheels
+      - build_windows_wheel
     runs-on: ubuntu-latest
 
     environment:
