+++ .github/workflows/python-package.yml
@@ -146,7 +146,7 @@ jobs:
 
     environment:
       name: testpypi
-      url: https://test.pypi.org/p/<package-name>
+      url: https://test.pypi.org/p/esat
 
     permissions:
       id-token: write  # IMPORTANT: mandatory for trusted publishing
