+++ .github/workflows/python-package.yml
@@ -40,7 +40,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: python-package-distributions
-          path: py_dist/
+          path: dist/
 
   build_windows_wheel:
     name: Build wheels on windows-latest
@@ -69,7 +69,7 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: python-package-distributions
-          path: py_dist/
+          path: dist/
 
 #  publish-to-pypi:
 #    name: >-
@@ -110,13 +110,13 @@ jobs:
 #        uses: actions/download-artifact@v4
 #        with:
 #          name: python-package-distributions
-#          path: py_dist/
+#          path: dist/
 #      - name: Sign the dists with Sigstore
 #        uses: sigstore/gh-action-sigstore-python@v3.0.0
 #        with:
 #          inputs: >-
-#            ./py_dist/*.tar.gz
-#            ./py_dist/*.whl
+#            ./dist/*.tar.gz
+#            ./dist/*.whl
 #      - name: Create GitHub Release
 #        env:
 #          GITHUB_TOKEN: ${{ github.token }}
@@ -133,7 +133,7 @@ jobs:
 #        # sigstore-produced signatures and certificates.
 #        run: >-
 #          gh release upload
-#          '${{ github.ref_name }}' py_dist/**
+#          '${{ github.ref_name }}' dist/**
 #          --repo '${{ github.repository }}'
 
 
@@ -156,7 +156,7 @@ jobs:
         uses: actions/download-artifact@v4
         with:
           name: python-package-distributions
-          path: py_dist/
+          path: dist/
       - name: Publish distribution 📦 to TestPyPI
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
