### 2024-12-02 5114e27
updated github action triggers, updated pyproject.toml for test pypi
