### 2024-12-02 634a793
switched to required manylinux build using cibuildwheel for pypa
