### 2024-12-02 d3aafbf
added configuration details for cibuildwheel on github
