+++ pyproject.toml
@@ -35,6 +35,14 @@ Issues = "https://github.com/quanted/esat/issues"
 package-dir = {"esat" = "esat", "esat_eval" = "eval"}
 include-package-data = true
 
+[tool.cibuildwheel]
+skip = "*-musllinux_i686" # No musllinux-i686 rust compiler available.
+linux.before-all = "curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --profile minimal"
+linux.environment = { PATH="$HOME/.cargo/bin:$PATH" }
+macos.before-all = "curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --profile minimal"
+windows.before-all = "rustup target add aarch64-pc-windows-msvc i586-pc-windows-msvc i686-pc-windows-msvc x86_64-pc-windows-msvc"
+windows.archs = "all"
+
 [tool.setuptools.dynamic]
 dependencies = {file = ["requirements.txt"]}
 
@@ -49,3 +57,5 @@ target = "esat_rust"
 path = "Cargo.toml"
 binding = "PyO3"
 debug = false
+
+[tool.
