### 2024-12-02 e8a4685
fixed pyproject.toml typo
