### 2024-12-02 e3f2ff0
added env variables for ciwheelbuild workflow errors
