### 2024-12-02 73e9878
added env variables for ciwheelbuild workflow errors
