+++ .github/workflows/package-publish.yml
@@ -8,7 +8,7 @@ on:
       - dev
 
 env:
-  MACOSX_DEPLOYMENT_TARGET=10.12
+  MACOSX_DEPLOYMENT_TARGET: "10.12"
 
 jobs:
   build_wheels:
