+++ pyproject.toml
@@ -36,7 +36,7 @@ package-dir = {"esat" = "esat", "esat_eval" = "eval"}
 include-package-data = true
 
 [tool.cibuildwheel]
-skip = "cp36-* cp37-* cp-38* pp37-* *-manylinux_i686 *_ppc64le *_s390x *_universal2 *-musllinux_i686"
+skip = "cp36-* cp37-* cp-38* cp-313* pp37-* *-manylinux_i686 *_ppc64le *_s390x *_universal2 *-musllinux_i686"
 linux.before-all = "curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --profile minimal"
 linux.environment = { PATH="$HOME/.cargo/bin:$PATH" }
 macos.before-all = "curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y --profile minimal"
