+++ .github/workflows/package-publish.yml
@@ -12,7 +12,7 @@ env:
 
 jobs:
   build_wheels:
-    name: Build wheels on ${{ matrix.os }}
+    name: Build wheels on ${{ matrix.os }} {{ matrix.python_version }}
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
