### 2024-12-02 b7c09d1
attempt 2 for preventing 3.13 compilation
