+++ .github/workflows/package-publish.yml
@@ -12,11 +12,11 @@ env:
 
 jobs:
   build_wheels:
-    name: Build wheels on ${{ matrix.os }} {{ matrix.python_version }}
+    name: Build wheels on ${{ matrix.os }} ${{ matrix.python_version }}
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        os: [ubuntu-latest, windows-latest, macos-13, macos-14]
+        os: [ubuntu-latest, windows-latest, macos-13, macos-latest]
         python_version: ["3.10", "3.11", "3.12"]
     steps:
       - uses: actions/checkout@v4
