### 2024-12-02 a21d2d3
added specific python target versions to github action
