### 2024-12-03 #7316 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
