### 2024-12-09 0919859
paper updates to maintain captialization, workflow build updates
