+++ .github/workflows/python-package.yml
@@ -3,7 +3,7 @@ name: Build and Publish Artifacts
 on:
   push:
     branches:
-      - main
+      - dev
 
 jobs:
   build_wheels:
@@ -11,49 +11,76 @@ jobs:
     runs-on: ${{ matrix.os }}
     strategy:
       matrix:
-        # macos-13 is an intel runner, macos-14 is apple silicon
-        os: [ubuntu-latest, macos-13, macos-14]
-        python_version: ["3.10", "3.11", "3.12"]
+        os: [ ubuntu-latest, windows-latest, macos-13, macos-latest ]
     steps:
       - uses: actions/checkout@v4
       - uses: actions/setup-python@v5
         with:
-          python-version: ${{ matrix.python_version }}
+          python-version: "3.12"
+      - name: Install cibuildwheel
+        run: python -m pip install cibuildwheel==2.22.0
       - name: Install setuptools
         run: |
-            python -m pip install --upgrade pip
-            curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain=stable
-            echo "PATH=${PATH}:${HOME}/.cargo/bin" >> "$GITHUB_ENV"
-            rustc --version
-            python -m pip install build setuptools setuptools-rust
+          python -m pip install --upgrade pip
+          curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain=stable
+          echo "PATH=${PATH}:${HOME}/.cargo/bin" >> "$GITHUB_ENV"
+          rustc --version
+          python -m pip install build setuptools setuptools-rust
       - name: Build wheels
-        run:
-            python -m build --outdir wheelhouse
+        run: python -m cibuildwheel --output-dir wheelhouse
       - uses: actions/upload-artifact@v4
         with:
           name: esat-wheels-${{ matrix.os }}-py${{matrix.python_version}}
           path: ./wheelhouse/*.whl
 
-  build_windows_wheel:
-    name: Build wheels on windows-latest
-    runs-on: windows-latest
-    strategy:
-      matrix:
-        python_version: [ "3.10", "3.11", "3.12" ]
-    steps:
-      - uses: actions/checkout@v4
-      - uses: actions/setup-python@v5
-        with:
-          python-version: ${{ matrix.python_version }}
-      - name: Install setuptools
-        run: |
-          python -m pip install --upgrade pip
-          curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain=stable
-          echo "${HOME}/.cargo/bin" | Out-File -FilePath $env:GITHUB_PATH -Encoding utf8 -Append
-          python -m pip install build setuptools setuptools-rust
-      - name: Build wheels
-        run: python -m build --outdir wheelhouse
-      - uses: actions/upload-artifact@v4
-        with:
-          name: esat-wheels-windows-py${{matrix.python_version}}
-          path: ./wheelhouse/*.whl
+#  build_wheels:
+#    name: Build wheels on ${{ matrix.os }}
+#    runs-on: ${{ matrix.os }}
+#    strategy:
+#      matrix:
+#        # macos-13 is an intel runner, macos-14 is apple silicon
+#        os: [ubuntu-latest, macos-13, macos-14]
+#        python_version: ["3.10", "3.11", "3.12"]
+#    steps:
+#      - uses: actions/checkout@v4
+#      - uses: actions/setup-python@v5
+#        with:
+#          python-version: ${{ matrix.python_version }}
+#      - name: Install setuptools
+#        run: |
+#            python -m pip install --upgrade pip
+#            curl https://sh.rustup.rs -sSf | sh -s -- -y --default-toolchain=stable
[…]
