### 2024-12-09 08a13e2
paper updates to maintain captialization, workflow build updates
