+++ _sources/docs/esat.rst
@@ -1,0 +1,57 @@
+esat package
+============
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   esat.cli
+   esat.data
+   esat.error
+   esat.model
+   esat.rotational
+
+Submodules
+----------
+
+esat.configs module
+-------------------
+
+.. automodule:: esat.configs
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.metrics module
+-------------------
+
+.. automodule:: esat.metrics
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.utils module
+-----------------
+
+.. automodule:: esat.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+esat.estimator module
+---------------------
+
+.. automodule:: esat.estimator
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: esat
+   :members:
+   :undoc-members:
+   :show-inheritance:
