+++ _sources/README.md
@@ -1,0 +1,193 @@
+# Environmental Source Apportionment Toolkit (ESAT)
+Last Update: 12-09-2024
+
+## Table of Contents
+ - [Description](#description)
+ - [Quick Start](#quick-start)
+   - [Installation](#installation) 
+   - [Example Code](#example-code)
+ - [Background](#background)
+    - [ESAT Predecessor](#esat-predecessor)
+    - [Features](#features)
+    - [Notebooks](#notebooks)
+ - [Development](#development)
+    - [Requirements](#requirements)
+    - [Rust Compiling](#rust-compiling)
+    - [Creating Docs](#creating-docs)
+    - [Building Wheels and Compiling Rust](#building-wheels-and-compiling-rust)
+    - [Code Tests](#code-tests)
+    - [Community Contributions](#community-contributions)
+
+## Description
+The Environmental Source Apportionment Toolkit (ESAT) is an open-source software package that provides API and CLI 
+functionality to create source apportionment workflows specifically targeting environmental datasets. Source 
+apportionment in environment science is the process of mathematically estimating the profiles and contributions of 
+multiple sources in some dataset, and in the case of ESAT, while considering data uncertainty. There are many potential 
+use cases for source apportionment in environmental science research, such as in the fields of air quality, water 
+quality and potentially many others.
+
+The ESAT toolkit is written in Python and Rust, and uses common packages such as numpy, scipy and pandas for data 
+processing. The source apportionment algorithms provided in ESAT include two variants of non-negative matrix 
+factorization (NMF), both of which have been written in Rust and contained within the python package. A collection of 
+data processing and visualization features are included for data and model analytics. The ESAT package includes a 
+synthetic data generator and comparison tools to evaluate ESAT model outputs.
+
+## Quick Start
+
+ESAT supports python version 3.10, 3.11, and 3.12. As the python package contains compiled code there are OS and python
+specific versions, supporting Mac Intel, Mac M1+, Linux, and Windows.
+
+### Documentation
+The Python API and CLI documentation can be found at the GitHub ESAT IO site: [https://quanted.github.io/esat/](https://quanted.github.io/esat/)
+
+### Installation
+The ESAT python package contains all compiled code and required dependencies and can be installed using pip
+```bash
+pip install esat
+```
+which will install the latest version that supports and is available for your python version and OS.
+
+Development versions of ESAT can be found on the GitHub actions page, for logged-in users, 
+under the 'Build and Publish Wheel' workflow. The latest version of the package will be available as an artifact for 
+download in the 'Artifacts' section of the completed workflow. There wheel files can be found for specific versions
+of python and supported operating systems. 
+
+If an error message appears during installation stating that the 
+package is not supported check that the correct OS and python version are being installed for that system. The python 
+wheels can be installed directly using 
+```bash
+pip install <wheel file name>
+```
+The esat python package is recommended to be installed in its own dedicated python virtual environment or conda environment.
+
+To run the jupyter notebooks, install jupyterlab into the esat python environmental
+```bash
+pip install jupyterlab
+```
+
+### Example Code
+Jupyter notebooks containing complete code examples, using sample datasets, are available for the 
+[source apportionment workflow](notebooks/epa_esat_workflow_01.ipynb) and the [simulator workflow](notebooks/epa_esat_simulator_01.ipynb). 
+
+## Background
+
+### ESAT Predecessor
+A widely used application used for environmental source apportionment is the EPA's Positive Matrix Factorization version
+5 (PMF5), which is a broadly used tool with an international user community. The PMF5 application is a mathematical tool
+that processes a dataset of feature concentrations across many samples (and concentration uncertainties) to estimate a 
+set of source profiles and their contributions. PMF5 can be used on a wide range of environmental data and is a powerful
+tool for estimating source apportionment: 
[…]
