### 2024-12-09 #7316 @deronsmith
@kthyng Thank you for catching those reference formatting inconsistencies. I have updated the bib to maintain all capitalizations, based upon source formatting and proper nouns, with merge [b2f7cc7](https://github.com/quanted/esat/commit/b2f7cc78bac0682111e5d8b75b456ff1c3804a45).
