### 2024-12-10 #7316 @kthyng
Congratulations on your new publication @deronsmith! Many thanks to editor @mengqi-z and to reviewers @gutabeshu,  @ifoxfoot, and @niravlekinwala for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
